tools/telegram-admin/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
python-dotenv==1.0.0
jinja2==3.1.3
python-multipart==0.0.9
